+++ src/php/Allegro/Protobuf/Compiler/DescriptorInterface.php
@@ -1,5 +1,5 @@
 <?php
-namespace ProtobufCompiler;
+namespace Allegro\Protobuf\Compiler;
 
 interface DescriptorInterface
 {
