+++ src/php/Allegro/Protobuf/Compiler/FieldDescriptor.php
@@ -1,7 +1,8 @@
 <?php
-namespace ProtobufCompiler;
+namespace Allegro\Protobuf\Compiler;
 
-require_once 'pb_proto_descriptor.php';
+use Google\Protobuf\Compiler\FieldDescriptorProto_Label;
+use Google\Protobuf\Compiler\FieldDescriptorProto_Type;
 
 /**
  * Describes field
@@ -9,42 +10,42 @@ require_once 'pb_proto_descriptor.php';
 class FieldDescriptor
 {
     private static $_scalarInternalTypesByProtobufType = array(
-        \FieldDescriptorProto_Type::TYPE_DOUBLE   => \ProtobufMessage::PB_TYPE_DOUBLE,
-        \FieldDescriptorProto_Type::TYPE_FLOAT    => \ProtobufMessage::PB_TYPE_FLOAT,
-        \FieldDescriptorProto_Type::TYPE_INT32    => \ProtobufMessage::PB_TYPE_INT,
-        \FieldDescriptorProto_Type::TYPE_INT64    => \ProtobufMessage::PB_TYPE_INT,
-        \FieldDescriptorProto_Type::TYPE_UINT32   => \ProtobufMessage::PB_TYPE_INT,
-        \FieldDescriptorProto_Type::TYPE_UINT64   => \ProtobufMessage::PB_TYPE_INT,
-        \FieldDescriptorProto_Type::TYPE_SINT32   => \ProtobufMessage::PB_TYPE_SIGNED_INT,
-        \FieldDescriptorProto_Type::TYPE_SINT64   => \ProtobufMessage::PB_TYPE_SIGNED_INT,
-        \FieldDescriptorProto_Type::TYPE_FIXED32  => \ProtobufMessage::PB_TYPE_FIXED32,
-        \FieldDescriptorProto_Type::TYPE_FIXED64  => \ProtobufMessage::PB_TYPE_FIXED64,
-        \FieldDescriptorProto_Type::TYPE_SFIXED32 => \ProtobufMessage::PB_TYPE_FIXED32,
-        \FieldDescriptorProto_Type::TYPE_SFIXED64 => \ProtobufMessage::PB_TYPE_FIXED64,
-        \FieldDescriptorProto_Type::TYPE_BOOL     => \ProtobufMessage::PB_TYPE_BOOL,
-        \FieldDescriptorProto_Type::TYPE_STRING   => \ProtobufMessage::PB_TYPE_STRING,
-        \FieldDescriptorProto_Type::TYPE_BYTES    => \ProtobufMessage::PB_TYPE_STRING,
-        \FieldDescriptorProto_Type::TYPE_ENUM     => \ProtobufMessage::PB_TYPE_INT,
+        FieldDescriptorProto_Type::TYPE_DOUBLE   => \ProtobufMessage::PB_TYPE_DOUBLE,
+        FieldDescriptorProto_Type::TYPE_FLOAT    => \ProtobufMessage::PB_TYPE_FLOAT,
+        FieldDescriptorProto_Type::TYPE_INT32    => \ProtobufMessage::PB_TYPE_INT,
+        FieldDescriptorProto_Type::TYPE_INT64    => \ProtobufMessage::PB_TYPE_INT,
+        FieldDescriptorProto_Type::TYPE_UINT32   => \ProtobufMessage::PB_TYPE_INT,
+        FieldDescriptorProto_Type::TYPE_UINT64   => \ProtobufMessage::PB_TYPE_INT,
+        FieldDescriptorProto_Type::TYPE_SINT32   => \ProtobufMessage::PB_TYPE_SIGNED_INT,
+        FieldDescriptorProto_Type::TYPE_SINT64   => \ProtobufMessage::PB_TYPE_SIGNED_INT,
+        FieldDescriptorProto_Type::TYPE_FIXED32  => \ProtobufMessage::PB_TYPE_FIXED32,
+        FieldDescriptorProto_Type::TYPE_FIXED64  => \ProtobufMessage::PB_TYPE_FIXED64,
+        FieldDescriptorProto_Type::TYPE_SFIXED32 => \ProtobufMessage::PB_TYPE_FIXED32,
+        FieldDescriptorProto_Type::TYPE_SFIXED64 => \ProtobufMessage::PB_TYPE_FIXED64,
+        FieldDescriptorProto_Type::TYPE_BOOL     => \ProtobufMessage::PB_TYPE_BOOL,
+        FieldDescriptorProto_Type::TYPE_STRING   => \ProtobufMessage::PB_TYPE_STRING,
+        FieldDescriptorProto_Type::TYPE_BYTES    => \ProtobufMessage::PB_TYPE_STRING,
+        FieldDescriptorProto_Type::TYPE_ENUM     => \ProtobufMessage::PB_TYPE_INT,
     );
 
     private static $_phpTypesByProtobufType = array(
-        \FieldDescriptorProto_Type::TYPE_DOUBLE   => 'double',
-        \FieldDescriptorProto_Type::TYPE_FLOAT    => 'double',
-        \FieldDescriptorProto_Type::TYPE_INT32    => 'integer',
-        \FieldDescriptorProto_Type::TYPE_INT64    => 'integer',
-        \FieldDescriptorProto_Type::TYPE_UINT32   => 'integer',
-        \FieldDescriptorProto_Type::TYPE_UINT64   => 'integer',
-        \FieldDescriptorProto_Type::TYPE_SINT32   => 'integer',
-        \FieldDescriptorProto_Type::TYPE_SINT64   => 'integer',
-        \FieldDescriptorProto_Type::TYPE_FIXED32  => 'integer',
-        \FieldDescriptorProto_Type::TYPE_FIXED64  => 'integer',
-        \FieldDescriptorProto_Type::TYPE_SFIXED32 => 'integer',
-        \FieldDescriptorProto_Type::TYPE_SFIXED64 => 'integer',
-        \FieldDescriptorProto_Type::TYPE_BOOL     => 'boolean',
-        \FieldDescriptorProto_Type::TYPE_STRING   => 'string',
-        \FieldDescriptorProto_Type::TYPE_BYTES    => 'string',
-        \FieldDescriptorProto_Type::TYPE_ENUM     => 'integer',
-        \FieldDescriptorProto_Type::TYPE_MESSAGE  => 'object'
+        FieldDescriptorProto_Type::TYPE_DOUBLE   => 'double',
+        FieldDescriptorProto_Type::TYPE_FLOAT    => 'double',
+        FieldDescriptorProto_Type::TYPE_INT32    => 'integer',
+        FieldDescriptorProto_Type::TYPE_INT64    => 'integer',
+        FieldDescriptorProto_Type::TYPE_UINT32   => 'integer',
+        FieldDescriptorProto_Type::TYPE_UINT64   => 'integer',
+        FieldDescriptorProto_Type::TYPE_SINT32   => 'integer',
+        FieldDescriptorProto_Type::TYPE_SINT64   => 'integer',
+        FieldDescriptorProto_Type::TYPE_FIXED32  => 'integer',
+        FieldDescriptorProto_Type::TYPE_FIXED64  => 'integer',
+        FieldDescriptorProto_Type::TYPE_SFIXED32 => 'integer',
+        FieldDescriptorProto_Type::TYPE_SFIXED64 => 'integer',
+        FieldDescriptorProto_Type::TYPE_BOOL     => 'boolean',
+        FieldDescriptorProto_Type::TYPE_STRING   => 'string',
+        FieldDescriptorProto_Type::TYPE_BYTES    => 'string',
+        FieldDescriptorProto_Type::TYPE_ENUM     => 'integer',
+        FieldDescriptorProto_Type::TYPE_MESSAGE  => 'object'
     );
 
     private $_default;
@@ -178,7 +179,7 @@ class FieldDescriptor
      */
     public function isRepeated()
     {
-        return $this->_label == \FieldDescriptorProto_Label::LABEL_REPEATED;
+        return $this->_label == FieldDescriptorProto_Label::LABEL_REPEATED;
     }
 
     /**
@@ -188,7 +189,7 @@ class FieldDescriptor
      */
     public function isRequired()
     {
-        return $this->_label == \FieldDescriptorProto_Label::LABEL_REQUIRED;
+        return $this->_label == FieldDescriptorProto_Label::LABEL_REQUIRED;
     }
 
     /**
@@ -198,7 +199,7 @@ class FieldDescriptor
      */
     public function isOptional()
     {
-        return $this->_label == \FieldDescriptorProto_Label::LABEL_OPTIONAL;
+        return $this->_label == FieldDescriptorProto_Label::LABEL_OPTIONAL;
     }
 
     /**
