+++ src/php/Allegro/Protobuf/Compiler/FileDescriptor.php
@@ -1,5 +1,5 @@
 <?php
-namespace ProtobufCompiler;
+namespace Allegro\Protobuf\Compiler;
 
 /**
  * Describes file
